new fruit spin component
