--- app/lab/scout-spin/components/badge.tsx
@@ -1,76 +1,0 @@
-export function Badge() {
-  return (
-    <svg width="59" height="59" viewBox="0 0 59 59" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <g clipPath="url(#clip0_1009_505)">
-        <path
-          d="M17.2042 11.0498C19.8235 6.13716 26.0533 4.46788 30.7781 7.41274C31.6645 7.96523 32.7468 8.10758 33.7459 7.80342C39.072 6.18163 44.6583 9.40692 45.9168 14.8303C46.153 15.8477 46.8175 16.7138 47.7391 17.2052C52.6518 19.8245 54.3211 26.0543 51.3762 30.7791C50.8238 31.6655 50.6814 32.7478 50.9856 33.747C52.6073 39.073 49.3821 44.6594 43.9587 45.9179C42.9413 46.154 42.0752 46.8185 41.5838 47.7401C38.9645 52.6528 32.7346 54.3221 28.0099 51.3772C27.1235 50.8248 26.0412 50.6824 25.042 50.9866C19.716 52.6084 14.1296 49.3831 12.8711 43.9597C12.6349 42.9423 11.9705 42.0762 11.0488 41.5848C6.13618 38.9655 4.4669 32.7356 7.41176 28.0109C7.96424 27.1245 8.1066 26.0422 7.80244 25.043C6.18065 19.717 9.40594 14.1306 14.8293 12.8721C15.8467 12.6359 16.7128 11.9715 17.2042 11.0498Z"
-          fill="#F4BE97"
-          stroke="white"
-          strokeWidth="6"
-        />
-        <mask
-          id="mask0_1009_505"
-          style={{ maskType: 'luminance' }}
-          maskUnits="userSpaceOnUse"
-          x="15"
-          y="16"
-          width="29"
-          height="27"
-        >
-          <path
-            d="M38.3966 16.6321L15.2144 22.8438L20.3907 42.1623L43.573 35.9506L38.3966 16.6321Z"
-            fill="white"
-          />
-        </mask>
-        <g mask="url(#mask0_1009_505)">
-          <path
-            d="M19.1338 33.4388C18.8333 32.3175 18.6831 31.7569 18.6025 31.2724C17.8194 26.5669 20.4275 21.9793 24.8319 20.315C25.2854 20.1436 25.8387 19.9954 26.9452 19.6989C28.0517 19.4024 28.6049 19.2542 29.0834 19.1758C33.7298 18.4149 38.2823 21.0839 39.9569 25.5505C40.1293 26.0104 40.2795 26.5711 40.58 27.6923L40.8537 28.714C41.2006 30.0085 41.374 30.6557 41.421 31.2215C41.636 33.8083 40.2481 36.2493 37.9347 37.3538C37.4287 37.5954 36.79 37.7665 35.5126 38.1088L28.7305 39.9261C27.4531 40.2683 26.8144 40.4395 26.2554 40.4833C23.6996 40.6835 21.2772 39.2634 20.17 36.9157C19.9278 36.4022 19.7544 35.755 19.4075 34.4605L19.1338 33.4388Z"
-            fill="#D58043"
-          />
-          <path
-            d="M21.2469 29.5612C19.7735 30.4184 17.7452 29.6579 16.7165 27.8625C15.6878 26.067 16.0483 23.9166 17.5217 23.0594C18.9951 22.2021 21.0234 22.9627 22.0521 24.7581C23.0808 26.5535 22.7203 28.7039 21.2469 29.5612Z"
-            fill="#D58043"
-          />
-          <path
-            d="M36.5081 17.9746C38.2128 17.9803 39.6002 19.6624 39.607 21.7317C39.6139 23.8009 38.2375 25.4737 36.5329 25.468C34.8283 25.4623 33.4409 23.7802 33.4341 21.711C33.4272 19.6418 34.8035 17.9689 36.5081 17.9746Z"
-            fill="#D58043"
-          />
-          <path
-            d="M23.3635 27.7765C23.0955 26.7762 23.6784 25.7509 24.6654 25.4865C25.6525 25.222 26.6699 25.8185 26.9379 26.8187L27.7546 29.8667C27.9289 30.5172 27.5498 31.1841 26.9078 31.3561L25.6583 31.6909C25.0163 31.8629 24.3545 31.475 24.1802 30.8244L23.3635 27.7765Z"
-            fill="#F4BE97"
-          />
-          <path
-            d="M30.5124 25.8604C30.2444 24.8602 30.8273 23.8349 31.8143 23.5705C32.8014 23.306 33.8188 23.9024 34.0868 24.9027L34.9035 27.9507C35.0778 28.6012 34.6987 29.2681 34.0567 29.4401L32.8072 29.7749C32.1652 29.9469 31.5035 29.559 31.3291 28.9084L30.5124 25.8604Z"
-            fill="#F4BE97"
-          />
-        </g>
-        <mask
-          id="mask1_1009_505"
-          style={{ maskType: 'alpha' }}
-          maskUnits="userSpaceOnUse"
-          x="4"
-          y="5"
-          width="49"
-          height="49"
-        >
-          <path
-            d="M35.1226 52.2954C47.9258 48.8648 55.5238 35.7047 52.0932 22.9015C48.6626 10.0983 35.5025 2.50032 22.6993 5.93092C9.89612 9.36152 2.29814 22.5216 5.72875 35.3248C9.15935 48.128 22.3194 55.726 35.1226 52.2954Z"
-            fill="#FFC362"
-          />
-        </mask>
-        <g mask="url(#mask1_1009_505)">
-          <path
-            d="M38.5762 -0.0876048L8.20117 52.5234L18.5935 58.5234L48.9685 5.9124L38.5762 -0.0876048Z"
-            fill="white"
-            fillOpacity="0.2"
-          />
-        </g>
-      </g>
-      <defs>
-        <clipPath id="clip0_1009_505">
-          <rect width="59" height="59" fill="white" />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-}
