--- app/lab/scout-spin/components/card-bg.tsx
@@ -1,54 +1,0 @@
-export function CardBackground() {
-  return (
-    <svg
-      width="282"
-      height="157"
-      viewBox="0 0 282 157"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g clipPath="url(#clip0_1006_456)">
-        <path
-          d="M12.7499 129.702C12.6905 116.352 12.6608 109.677 13.1901 104.055C18.3306 49.4564 61.383 6.25941 116.018 0.881877C121.644 0.328125 128.326 0.328125 141.69 0.328125C155.055 0.328125 161.737 0.328125 167.368 0.881877C222.051 6.25941 265.487 49.4564 271.113 104.055C271.693 109.677 271.722 116.352 271.782 129.702L271.836 141.865C271.904 157.277 271.939 164.983 270.789 171.407C265.531 200.781 242.623 223.766 213.241 229.149C206.815 230.326 199.1 230.326 183.672 230.326H101.755C86.3266 230.326 78.6123 230.326 72.1757 229.149C42.746 223.766 19.6335 200.781 14.114 171.407C12.9068 164.983 12.8726 157.277 12.804 141.865L12.7499 129.702Z"
-          stroke="#323232"
-          strokeOpacity="0.02"
-          strokeWidth="2"
-        />
-        <path
-          d="M48.0984 92.939C28.929 98.0758 8.30339 83.5906 2.02997 60.5851C-4.24346 37.5796 6.21087 14.7657 25.3803 9.62874C44.5499 4.49182 65.1753 18.9772 71.4488 41.9826C77.7223 64.9881 67.2677 87.8021 48.0984 92.939Z"
-          stroke="#323232"
-          strokeOpacity="0.02"
-          strokeWidth="2"
-        />
-        <path
-          d="M254.702 9.6307C273.918 14.7676 284.575 37.5816 278.506 60.5871C272.438 83.5926 251.941 98.0778 232.726 92.9409C213.51 87.8041 202.853 64.9901 208.922 41.9846C214.991 18.9791 235.487 4.49377 254.702 9.6307Z"
-          stroke="#323232"
-          strokeOpacity="0.02"
-          strokeWidth="2"
-        />
-        <path
-          d="M76.2038 79.3885C76.1509 67.48 85.7726 57.8262 97.6941 57.8262C109.616 57.8262 119.323 67.48 119.376 79.3885L119.537 115.676C119.572 123.422 113.314 129.7 105.56 129.7H90.4675C82.7137 129.7 76.3997 123.422 76.3653 115.676L76.2038 79.3885Z"
-          stroke="#323232"
-          strokeOpacity="0.02"
-          strokeWidth="2"
-        />
-        <path
-          d="M162.548 79.3885C162.495 67.48 172.116 57.8262 184.038 57.8262C195.96 57.8262 205.667 67.48 205.72 79.3885L205.882 115.676C205.916 123.422 199.658 129.7 191.904 129.7H176.812C169.058 129.7 162.744 123.422 162.71 115.676L162.548 79.3885Z"
-          stroke="#323232"
-          strokeOpacity="0.02"
-          strokeWidth="2"
-        />
-      </g>
-      <defs>
-        <clipPath id="clip0_1006_456">
-          <rect
-            width="280"
-            height="230"
-            fill="white"
-            transform="matrix(1 0 0.00444879 0.99999 0 0.328125)"
-          />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-}
