--- app/lab/scout-spin/components/coin-shadow.tsx
@@ -1,7 +1,0 @@
-export function CoinShadow() {
-  return (
-    <svg width="61" height="8" viewBox="0 0 61 8" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <ellipse cx="30.25" cy="4" rx="30" ry="4" fill="#B3B3B3" />
-    </svg>
-  );
-}
