--- app/lab/scout-spin/components/coin.tsx
@@ -1,74 +1,0 @@
-export function Coin() {
-  return (
-    <svg width="65" height="68" viewBox="0 0 65 68" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <path
-        d="M32.75 0C50.4231 0 64.75 14.3269 64.75 32C64.75 33.3547 64.75 33 64.75 34C64.75 38.2023 63.9226 44.3637 62.3145 48.2461C60.7063 52.1285 58.3484 55.6555 55.377 58.627C52.4055 61.5984 48.8785 63.9563 44.9961 65.5645C41.1137 67.1726 36.9523 68 32.75 68C28.5477 68 24.3863 67.1726 20.5039 65.5645C16.6215 63.9563 13.0945 61.5984 10.123 58.627C7.15159 55.6555 4.7937 52.1285 3.18555 48.2461C1.5774 44.3637 0.75 38.2023 0.75 34C0.750115 32.5 0.75 33.3547 0.75 32C0.75 14.3269 15.0769 0 32.75 0Z"
-        fill="#FA9F1E"
-      />
-      <circle cx="32.75" cy="32" r="32" fill="#FFC362" />
-      <mask
-        id="mask0_987_284"
-        style={{ maskType: 'alpha' }}
-        maskUnits="userSpaceOnUse"
-        x="0"
-        y="0"
-        width="65"
-        height="64"
-      >
-        <circle cx="32.75" cy="32" r="32" fill="#FFC362" />
-      </mask>
-      <g mask="url(#mask0_987_284)">
-        <rect
-          x="-2"
-          y="55"
-          width="81"
-          height="16"
-          transform="rotate(-45 -2 55)"
-          fill="white"
-          fillOpacity="0.16"
-        />
-      </g>
-      <g clipPath="url(#clip0_987_284)">
-        <path
-          d="M33.1914 21C34.7036 21 35.4597 20.9999 36.0967 21.0625C42.2823 21.6704 47.1768 26.5536 47.7861 32.7256C47.8489 33.3611 47.8486 34.1159 47.8486 35.625V37C47.8486 38.7422 47.8482 39.6136 47.7148 40.3398C47.1049 43.6601 44.5014 46.2585 41.1738 46.8672C40.446 47.0003 39.5722 47 37.8262 47H28.5557C26.8098 47 25.9368 47.0003 25.209 46.8672C21.8812 46.2587 19.2769 43.6603 18.667 40.3398C18.5336 39.6136 18.5342 38.7422 18.5342 37V35.625C18.5342 34.1159 18.5339 33.3611 18.5967 32.7256C19.206 26.5537 24.0997 21.6706 30.2852 21.0625C30.9221 20.9999 31.679 21 33.1914 21ZM28.3057 27.5C26.9565 27.5 25.8623 28.5913 25.8623 29.9375V34.04C25.8625 34.9153 26.5739 35.6249 27.4512 35.625H29.1592C30.0365 35.625 30.7478 34.9154 30.748 34.04V29.9375C30.748 28.5914 29.6547 27.5001 28.3057 27.5ZM38.0771 27.5C36.728 27.5 35.6348 28.5913 35.6348 29.9375V34.04C35.635 34.9154 36.3462 35.625 37.2236 35.625H38.9316C39.8088 35.6247 40.5203 34.9152 40.5205 34.04V29.9375C40.5205 28.5913 39.4263 27.5 38.0771 27.5Z"
-          fill="#FA9F1E"
-        />
-        <path
-          d="M22.5529 31.4684C20.3809 32.0491 18.054 30.4116 17.3557 27.8109C16.6573 25.2103 17.8519 22.6313 20.0239 22.0506C22.1959 21.4699 24.5227 23.1074 25.2211 25.708C25.9195 28.3087 24.7249 30.8877 22.5529 31.4684Z"
-          fill="#FA9F1E"
-        />
-        <path
-          d="M45.976 22.0525C48.148 22.6332 49.3426 25.2122 48.6442 27.8129C47.9459 30.4135 45.6189 32.051 43.447 31.4703C41.275 30.8896 40.0804 28.3106 40.7788 25.71C41.4771 23.1093 43.804 21.4718 45.976 22.0525Z"
-          fill="#FA9F1E"
-        />
-      </g>
-      <mask
-        id="mask1_987_284"
-        style={{ maskType: 'alpha' }}
-        maskUnits="userSpaceOnUse"
-        x="0"
-        y="0"
-        width="65"
-        height="64"
-      >
-        <circle cx="32.75" cy="32" r="32" fill="#FFC362" />
-      </mask>
-      <g mask="url(#mask1_987_284)">
-        <rect
-          x="-2"
-          y="55"
-          width="81"
-          height="16"
-          transform="rotate(-45 -2 55)"
-          fill="white"
-          fillOpacity="0.08"
-        />
-      </g>
-      <defs>
-        <clipPath id="clip0_987_284">
-          <rect width="31.6875" height="26" fill="white" transform="translate(17.1562 21)" />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-}
