--- app/lab/scout-spin/components/icons.tsx
@@ -1,119 +1,0 @@
-import { JSX } from 'react';
-
-export function CentralArrowPathRightFilledOnStroke2Radius3(props: JSX.IntrinsicElements['svg']) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-    >
-      <path
-        d="M15.4603 4.41523C14.4833 3.60108 13 4.2958 13 5.56756V8.00004H6C3.79086 8.00004 2 9.7909 2 12C2 14.2092 3.79086 16 6 16H13V18.4325C13 19.7043 14.4833 20.399 15.4603 19.5849L22.2574 13.9206C23.4568 12.9211 23.4568 11.079 22.2574 10.0795L15.4603 4.41523Z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-}
-
-export function CentralCursorClickFilledOnStroke2Radius3(props: JSX.IntrinsicElements['svg']) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-    >
-      <path
-        d="M11 2C11.5523 2 12 2.44772 12 3V4.5C12 5.05228 11.5523 5.5 11 5.5C10.4477 5.5 10 5.05228 10 4.5V3C10 2.44772 10.4477 2 11 2Z"
-        fill="currentColor"
-      />
-      <path
-        d="M17.3634 4.63582C17.7539 5.02634 17.7539 5.65951 17.3634 6.05003L16.3027 7.11069C15.9122 7.50122 15.279 7.50122 14.8885 7.11069C14.498 6.72017 14.498 6.087 14.8885 5.69648L15.9491 4.63582C16.3397 4.24529 16.9728 4.24529 17.3634 4.63582Z"
-        fill="currentColor"
-      />
-      <path
-        d="M7.1114 16.303C7.50193 15.9124 7.50193 15.2793 7.1114 14.8887C6.72088 14.4982 6.08771 14.4982 5.69719 14.8887L4.63653 15.9494C4.24601 16.3399 4.24601 16.9731 4.63653 17.3636C5.02705 17.7541 5.66022 17.7541 6.05074 17.3636L7.1114 16.303Z"
-        fill="currentColor"
-      />
-      <path
-        d="M5.5 11C5.5 11.5523 5.05228 12 4.5 12H3C2.44772 12 2 11.5523 2 11C2 10.4477 2.44772 10 3 10H4.5C5.05228 10 5.5 10.4477 5.5 11Z"
-        fill="currentColor"
-      />
-      <path
-        d="M5.69704 7.1114C6.08756 7.50193 6.72073 7.50193 7.11125 7.1114C7.50178 6.72088 7.50178 6.08771 7.11125 5.69719L6.05059 4.63653C5.66007 4.24601 5.0269 4.24601 4.63638 4.63653C4.24585 5.02705 4.24585 5.66022 4.63638 6.05074L5.69704 7.1114Z"
-        fill="currentColor"
-      />
-      <path
-        d="M11.7286 9.45834C10.3036 8.89428 8.8933 10.3046 9.45736 11.7296L13.2007 21.1865C13.6561 22.337 15.2302 22.4724 15.8754 21.4166L17.9768 17.9778L21.4156 15.8763C22.4714 15.2311 22.336 13.6571 21.1855 13.2017L11.7286 9.45834Z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-}
-
-export function CentralChevronRightSmallFilledOnStroke2Radius3(
-  props: JSX.IntrinsicElements['svg'],
-) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M9.29289 7.29289C9.68342 6.90237 10.3166 6.90237 10.7071 7.29289L13.6464 10.2322C14.6228 11.2085 14.6228 12.7915 13.6464 13.7678L10.7071 16.7071C10.3166 17.0976 9.68342 17.0976 9.29289 16.7071C8.90237 16.3166 8.90237 15.6834 9.29289 15.2929L12.2322 12.3536C12.4275 12.1583 12.4275 11.8417 12.2322 11.6464L9.29289 8.70711C8.90237 8.31658 8.90237 7.68342 9.29289 7.29289Z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-}
-
-export function CentralCrossLargeFilledOnStroke2Radius3(props: JSX.IntrinsicElements['svg']) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M4.18934 4.18934C4.77513 3.60355 5.72487 3.60355 6.31066 4.18934L12 9.87868L17.6893 4.18934C18.2751 3.60355 19.2249 3.60355 19.8107 4.18934C20.3964 4.77513 20.3964 5.72487 19.8107 6.31066L14.1213 12L19.8107 17.6893C20.3964 18.2751 20.3964 19.2249 19.8107 19.8107C19.2249 20.3964 18.2751 20.3964 17.6893 19.8107L12 14.1213L6.31066 19.8107C5.72487 20.3964 4.77513 20.3964 4.18934 19.8107C3.60355 19.2249 3.60355 18.2751 4.18934 17.6893L9.87868 12L4.18934 6.31066C3.60355 5.72487 3.60355 4.77513 4.18934 4.18934Z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-}
-
-export function CentralHistoryFilledOnStroke2Radius3(props: JSX.IntrinsicElements['svg']) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M4 5.99932V5C4 4.44772 3.55228 4 3 4C2.44772 4 2 4.44772 2 5V9C2 9.55228 2.44772 10 3 10H7C7.55228 10 8 9.55228 8 9C8 8.44772 7.55228 8 7 8H5.06997C6.45395 5.60755 9.04015 4 12 4C16.4182 4 20 7.58172 20 12C20 16.4183 16.4182 20 12 20C8.51828 20 5.55363 17.7753 4.45492 14.6668C4.27088 14.146 3.69955 13.8731 3.17884 14.0572C2.65812 14.2412 2.3852 14.8125 2.56925 15.3332C3.94156 19.2159 7.64434 22 12 22C17.5228 22 22 17.5228 22 12C22 6.47715 17.5228 2 12 2C8.72763 2 5.82381 3.57164 4 5.99932ZM12 7C12.5523 7 13 7.44772 13 8V11.5858L15.7071 14.2929C16.0976 14.6834 16.0976 15.3166 15.7071 15.7071C15.3166 16.0976 14.6834 16.0976 14.2929 15.7071L11.2929 12.7071C11.1054 12.5196 11 12.2652 11 12V8C11 7.44772 11.4477 7 12 7Z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-}
