--- app/lab/scout-spin/components/sparkle.tsx
@@ -1,21 +1,0 @@
-export function Sparkle1() {
-  return (
-    <svg width="25" height="25" viewBox="0 0 25 25" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <path
-        d="M12.3523 1.7246C13.3236 0.142335 15.7392 0.622329 16.0318 2.45575L16.9975 8.50707C17.0895 9.08349 17.4287 9.59095 17.9262 9.89633L23.1486 13.1022C24.7308 14.0736 24.2508 16.4891 22.4174 16.7817L16.3661 17.7474C15.7897 17.8394 15.2822 18.1787 14.9768 18.6761L11.7709 23.8985C10.7996 25.4808 8.38405 25.0008 8.09145 23.1673L7.12572 17.116C7.03373 16.5396 6.69451 16.0321 6.19705 15.7268L0.974675 12.5208C-0.607595 11.5495 -0.127601 9.13398 1.70582 8.84138L7.75714 7.87565C8.33356 7.78366 8.84102 7.44444 9.1464 6.94698L12.3523 1.7246Z"
-        fill="#FFC362"
-      />
-    </svg>
-  );
-}
-
-export function Sparkle2() {
-  return (
-    <svg width="14" height="14" viewBox="0 0 14 14" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <path
-        d="M0.839032 4.77924C-0.0503162 3.70778 0.873224 2.10816 2.24581 2.34263L5.24177 2.85441C5.67331 2.92813 6.11551 2.80964 6.45238 2.53003L8.79107 0.588835C9.86253 -0.300513 11.4621 0.623028 11.2277 1.99561L10.7159 4.99158C10.6422 5.42311 10.7607 5.86532 11.0403 6.20218L12.9815 8.54087C13.8708 9.61233 12.9473 11.212 11.5747 10.9775L8.57874 10.4657C8.1472 10.392 7.70499 10.5105 7.36813 10.7901L5.02944 12.7313C3.95798 13.6206 2.35836 12.6971 2.59283 11.3245L3.10461 8.32854C3.17832 7.897 3.05983 7.4548 2.78023 7.11794L0.839032 4.77924Z"
-        fill="#FFC362"
-      />
-    </svg>
-  );
-}
