+++ components/sections/header.tsx
@@ -5,7 +5,15 @@ export default function Header() {
     <>
       <StickyHeader />
       <p className="text-tertiary mt-3 mb-24 text-sm">
-        Collections of ui components built by smintfy.
+        Collections of ui components built by smintfy. Check out the repository{' '}
+        <a
+          href="https://github.com/Smintfy/ui-snippet"
+          target="_blank"
+          className="hover:text-primary underline"
+        >
+          here
+        </a>
+        .
       </p>
     </>
   );
