+++ components/snippets/index.ts
@@ -1,0 +1,4 @@
+export { FamilyStyleOTP } from './family-style-otp';
+export { ImagePreview } from './image-preview';
+export { MultiStep } from './multi-step';
+export { FruitSpin } from './fruit-spin';
