--- components/ui/index.ts
@@ -1,4 +1,0 @@
-export { default as ComponentContainer } from './component-container';
-export { default as FamilyStyleOTP } from './family-style-otp';
-export { default as ImagePreview } from './image-preview';
-export { default as MultiStep } from './multi-step';
